# backend/api/gen/go/admin/service/v1/i_order_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/service/v1/i_order.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-shop/api/gen/go/order/service/v1"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationOrderServiceCount = "/admin.service.v1.OrderService/Count"
const OperationOrderServiceDelete = "/admin.service.v1.OrderService/Delete"
const OperationOrderServiceGet = "/admin.service.v1.OrderService/Get"
const OperationOrderServiceList = "/admin.service.v1.OrderService/List"
const OperationOrderServiceUpdate = "/admin.service.v1.OrderService/Update"

type OrderServiceHTTPServer interface {
	Count(context.Context, *v1.PagingRequest) (*v11.CountOrderResponse, error)
	Delete(context.Context, *v11.DeleteOrderRequest) (*emptypb.Empty, error)
	Get(context.Context, *v11.GetOrderRequest) (*v11.Order, error)
	List(context.Context, *v1.PagingRequest) (*v11.ListOrderResponse, error)
	Update(context.Context, *v11.UpdateOrderRequest) (*emptypb.Empty, error)
}

func RegisterOrderServiceHTTPServer(s *http.Server, srv OrderServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/mall/orders", _OrderService_List17_HTTP_Handler(srv))
	r.GET("/admin/v1/mall/orders/count", _OrderService_Count7_HTTP_Handler(srv))
	r.GET("/admin/v1/mall/orders/{id}", _OrderService_Get17_HTTP_Handler(srv))
	r.PUT("/admin/v1/mall/orders/{id}", _OrderService_Update12_HTTP_Handler(srv))
	r.DELETE("/admin/v1/mall/orders", _OrderService_Delete13_HTTP_Handler(srv))
}

func _OrderService_List17_HTTP_Handler(srv OrderServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListOrderResponse)
		return ctx.Result(200, reply)
	}
}

func _OrderService_Count7_HTTP_Handler(srv OrderServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderServiceCount)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Count(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.CountOrderResponse)
		return ctx.Result(200, reply)
	}
}

func _OrderService_Get17_HTTP_Handler(srv OrderServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetOrderRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetOrderRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Order)
		return ctx.Result(200, reply)
	}
}

func _OrderService_Update12_HTTP_Handler(srv OrderServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.UpdateOrderRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*v11.UpdateOrderRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _OrderService_Delete13_HTTP_Handler(srv OrderServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.DeleteOrderRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderServiceDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*v11.DeleteOrderRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type OrderServiceHTTPClient interface {
	Count(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.CountOrderResponse, err error)
	Delete(ctx context.Context, req *v11.DeleteOrderRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	Get(ctx context.Context, req *v11.GetOrderRequest, opts ...http.CallOption) (rsp *v11.Order, err error)
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListOrderResponse, err error)
	Update(ctx context.Context, req *v11.UpdateOrderRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
}

type OrderServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewOrderServiceHTTPClient(client *http.Client) OrderServiceHTTPClient {
	return &OrderServiceHTTPClientImpl{client}
}

func (c *OrderServiceHTTPClientImpl) Count(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.CountOrderResponse, error) {
	var out v11.CountOrderResponse
	pattern := "/admin/v1/mall/orders/count"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationOrderServiceCount))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *OrderServiceHTTPClientImpl) Delete(ctx context.Context, in *v11.DeleteOrderRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/mall/orders"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationOrderServiceDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *OrderServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetOrderRequest, opts ...http.CallOption) (*v11.Order, error) {
	var out v11.Order
	pattern := "/admin/v1/mall/orders/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationOrderServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *OrderServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListOrderResponse, error) {
	var out v11.ListOrderResponse
	pattern := "/admin/v1/mall/orders"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationOrderServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *OrderServiceHTTPClientImpl) Update(ctx context.Context, in *v11.UpdateOrderRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/mall/orders/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationOrderServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

const OperationOrderItemServiceBatchCreate = "/admin.service.v1.OrderItemService/BatchCreate"
const OperationOrderItemServiceCount = "/admin.service.v1.OrderItemService/Count"
const OperationOrderItemServiceCreate = "/admin.service.v1.OrderItemService/Create"
const OperationOrderItemServiceDelete = "/admin.service.v1.OrderItemService/Delete"
const OperationOrderItemServiceGet = "/admin.service.v1.OrderItemService/Get"
const OperationOrderItemServiceList = "/admin.service.v1.OrderItemService/List"
const OperationOrderItemServiceUpdate = "/admin.service.v1.OrderItemService/Update"

type OrderItemServiceHTTPServer interface {
	BatchCreate(context.Context, *v11.BatchCreateOrderItemsRequest) (*emptypb.Empty, error)
	Count(context.Context, *v1.PagingRequest) (*v11.CountOrderItemResponse, error)
	Create(context.Context, *v11.CreateOrderItemRequest) (*emptypb.Empty, error)
	Delete(context.Context, *v11.DeleteOrderItemRequest) (*emptypb.Empty, error)
	Get(context.Context, *v11.GetOrderItemRequest) (*v11.OrderItem, error)
	List(context.Context, *v1.PagingRequest) (*v11.ListOrderItemResponse, error)
	Update(context.Context, *v11.UpdateOrderItemRequest) (*emptypb.Empty, error)
}

func RegisterOrderItemServiceHTTPServer(s *http.Server, srv OrderItemServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/mall/order-items", _OrderItemService_List18_HTTP_Handler(srv))
	r.GET("/admin/v1/mall/order-items/count", _OrderItemService_Count8_HTTP_Handler(srv))
	r.GET("/admin/v1/mall/order-items/{id}", _OrderItemService_Get18_HTTP_Handler(srv))
	r.POST("/admin/v1/mall/order-items", _OrderItemService_Create13_HTTP_Handler(srv))
	r.POST("/admin/v1/mall/order-items/batch", _OrderItemService_BatchCreate5_HTTP_Handler(srv))
	r.PUT("/admin/v1/mall/order-items/{id}", _OrderItemService_Update13_HTTP_Handler(srv))
	r.DELETE("/admin/v1/mall/order-items", _OrderItemService_Delete14_HTTP_Handler(srv))
}

func _OrderItemService_List18_HTTP_Handler(srv OrderItemServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderItemServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListOrderItemResponse)
		return ctx.Result(200, reply)
	}
}

func _OrderItemService_Count8_HTTP_Handler(srv OrderItemServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderItemServiceCount)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Count(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.CountOrderItemResponse)
		return ctx.Result(200, reply)
	}
}

func _OrderItemService_Get18_HTTP_Handler(srv OrderItemServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetOrderItemRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderItemServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetOrderItemRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.OrderItem)
		return ctx.Result(200, reply)
	}
}

func _OrderItemService_Create13_HTTP_Handler(srv OrderItemServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.CreateOrderItemRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderItemServiceCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*v11.CreateOrderItemRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _OrderItemService_BatchCreate5_HTTP_Handler(srv OrderItemServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.BatchCreateOrderItemsRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderItemServiceBatchCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.BatchCreate(ctx, req.(*v11.BatchCreateOrderItemsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _OrderItemService_Update13_HTTP_Handler(srv OrderItemServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.UpdateOrderItemRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderItemServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*v11.UpdateOrderItemRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _OrderItemService_Delete14_HTTP_Handler(srv OrderItemServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.DeleteOrderItemRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOrderItemServiceDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*v11.DeleteOrderItemRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type OrderItemServiceHTTPClient interface {
	BatchCreate(ctx context.Context, req *v11.BatchCreateOrderItemsRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	Count(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.CountOrderItemResponse, err error)
	Create(ctx context.Context, req *v11.CreateOrderItemRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	Delete(ctx context.Context, req *v11.DeleteOrderItemRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	Get(ctx context.Context, req *v11.GetOrderItemRequest, opts ...http.CallOption) (rsp *v11.OrderItem, err error)
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListOrderItemResponse, err error)
	Update(ctx context.Context, req *v11.UpdateOrderItemRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
}

type OrderItemServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewOrderItemServiceHTTPClient(client *http.Client) OrderItemServiceHTTPClient {
	return &OrderItemServiceHTTPClientImpl{client}
}

func (c *OrderItemServiceHTTPClientImpl) BatchCreate(ctx context.Context, in *v11.BatchCreateOrderItemsRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/mall/order-items/batch"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationOrderItemServiceBatchCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *OrderItemServiceHTTPClientImpl) Count(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.CountOrderItemResponse, error) {
	var out v11.CountOrderItemResponse
	pattern := "/admin/v1/mall/order-items/count"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationOrderItemServiceCount))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *OrderItemServiceHTTPClientImpl) Create(ctx context.Context, in *v11.CreateOrderItemRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/mall/order-items"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationOrderItemServiceCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *OrderItemServiceHTTPClientImpl) Delete(ctx context.Context, in *v11.DeleteOrderItemRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/mall/order-items"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationOrderItemServiceDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *OrderItemServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetOrderItemRequest, opts ...http.CallOption) (*v11.OrderItem, error) {
	var out v11.OrderItem
	pattern := "/admin/v1/mall/order-items/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationOrderItemServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *OrderItemServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListOrderItemResponse, error) {
	var out v11.ListOrderItemResponse
	pattern := "/admin/v1/mall/order-items"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationOrderItemServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *OrderItemServiceHTTPClientImpl) Update(ctx context.Context, in *v11.UpdateOrderItemRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/mall/order-items/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationOrderItemServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}