# backend/app/app/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Content Hub App API
    description: GoWind Content Hub App API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-shop
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-shop/blob/master/LICENSE
    version: "1.0"
paths:
...
    /app/v1/mall/orders:
        get:
            tags:
                - OrderService
            operationId: OrderService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/order.service.v1.ListOrderResponse'
        post:
            tags:
                - OrderService
            operationId: OrderService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/order.service.v1.CreateOrderRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
...
    /app/v1/mall/orders/{id}:
        get:
            tags:
                - OrderService
            operationId: OrderService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: idempotencyKey
                  in: query
                  schema:
                    type: string
                - name: tenantId
                  in: query
                  description: 按 idempotency_key 查询时必须携带的租户ID（唯一索引为 tenant_id + idempotency_key）。
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/order.service.v1.Order'
        put:
            tags:
                - OrderService
            operationId: OrderService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/order.service.v1.UpdateOrderRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        order.service.v1.CreateOrderRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/order.service.v1.Order'
                couponId:
                    type: integer
                    description: 优惠券实例ID（核销指令，留空表示不使用）
                    format: uint32
...
        order.service.v1.ListOrderResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/order.service.v1.Order'
                total:
                    type: string
        order.service.v1.Order:
            type: object
            properties:
                id:
                    type: integer
                    description: 订单ID
                    format: uint32
                userId:
                    type: integer
                    description: 下单用户ID
                    format: uint32
                totalAmount:
                    type: string
                    description: 订单总金额（最小货币单位，分）
                currency:
                    type: string
                    description: 结算币种（ISO 4217）
                status:
                    enum:
                        - STATUS_UNSPECIFIED
                        - PENDING_PAYMENT
                        - PAID
                        - CANCELLED
                        - FULFILLED
                        - CLOSED
                    type: string
                    description: 订单状态
                    format: enum
                businessRefId:
                    type: string
                    description: 业务单号（对账主键，与支付流水的 business_ref_id 对齐）
                idempotencyKey:
                    type: string
                    description: 下单幂等键（客户端生成，租户内唯一）
                recipientName:
                    type: string
                    description: 收件人姓名
                recipientPhone:
                    type: string
                    description: 收件人电话
                shippingAddress:
                    type: string
                    description: 收货地址（结构化文本）
                originalAmount:
                    type: string
                    description: 订单折前总额（最小货币单位，分；各订单项 subtotal 之和，未扣抵扣；审计/对账用）
                discountAmount:
                    type: string
                    description: 优惠券抵扣额（最小货币单位，分；未用券为 0；审计/对账用）
                shippingFee:
                    type: string
                    description: 运费（最小货币单位，分；无运费规则时为 0；审计/对账用）
                taxAmount:
                    type: string
                    description: 税费（最小货币单位，分；无税率规则时为 0；审计/对账用）
                shippingRegion:
                    type: string
                    description: 收货地区代码（ISO 3166；运费/税率计算输入；审计/对账用）
                tenantId:
                    type: integer
                    description: 租户ID
                    format: uint32
                createdBy:
                    type: integer
                    description: 创建者用户ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者用户ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 订单
...
        order.service.v1.UpdateOrderRequest:
            type: object
            properties:
                id:
                    readOnly: true
                    type: integer
                    description: 订单ID
                    format: uint32
                data:
                    $ref: '#/components/schemas/order.service.v1.Order'
                updateMask:
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 允许缺失则插入
                expectedStatus:
                    type: array
                    items:
                        enum:
                            - STATUS_UNSPECIFIED
                            - PENDING_PAYMENT
                            - PAID
                            - CANCELLED
                            - FULFILLED
                            - CLOSED
                        type: string
                        format: enum
                    description: 状态前置条件白名单，订单当前状态须在其中更新才生效，否则返回 Conflict
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...